extract viewHelpers in cody
These are useful by other clients, not just the VS Code extension.
